Expose SerializerChain (#7)
This allows setting custom serializers so responses could be streamed,
instead of rendering the response into a string or a byte array and only
then starting to write it out to the client. Gives a significant
performance boost when used, for example, with a template engine that
supports outputting to a Writer or OutputStream.